refactor: split `Constraint.scala` into multiple files
