+++ main/src/ca/uwaterloo/flix/language/phase/monomorph2/Flow.scala
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2026 Simon Lykke Andersen
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package ca.uwaterloo.flix.language.phase.monomorph2
+
+/**
+  * A component-wise flow constraint.
+  * Read as: "The type-argument tuple `args` flows into the parameter slots of `dst`."
+  */
+case class Flow(args: List[MonoArg], dst: MonoVar)
